Connection interface 	add isConnected () Add DataUnserializer generic implementation class Move defer Recordset unserialization to this class PostgreSQL type map improvements

+++ tests/reference/PostgreSQL_Builder_create_Employees_connectionless_10.10.0.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS "ns_unittests"."Employees" (
-  "id" integer NOT NULL, 
+  "id" bigint NOT NULL, 
   "name" text, 
   "gender" text, 
   "salary" numeric(7, 2),
